# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/bookings/my-bookings": {
      "get": {
        "operationId": "BookingController_getMyBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get bookings made by current customer",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/provider-bookings": {
      "get": {
        "operationId": "BookingController_getProviderBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get bookings received by current provider",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "operationId": "BookingController_getBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get booking details by ID",
        "tags": [
          "Service Bookings"
        ]
      },
      "put": {
        "operationId": "BookingController_updateBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update booking details",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings": {
      "post": {
        "operationId": "BookingController_createBooking",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new booking",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/confirm": {
      "post": {
        "operationId": "BookingController_confirmBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a booking request",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/start": {
      "post": {
        "operationId": "BookingController_startBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark booking as started",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/complete": {
      "post": {
        "operationId": "BookingController_completeBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark booking as completed",
        "tags": [
          "Service Bookings"
        ]
      }
    },
...
    "/api/bookings/{id}/approve": {
      "post": {
        "operationId": "BookingController_approveDelivery",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve delivered work and unlock completion payment",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/correction": {
      "post": {
        "operationId": "BookingController_requestCorrection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCorrectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request a correction for delivered booking work",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingController_cancelBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a booking",
        "tags": [
          "Service Bookings"
        ]
      }
    },
...
    "/api/bookings/{id}/files": {
      "post": {
        "operationId": "BookingFileController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingFileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a workspace file for a booking",
        "tags": [
          "Service Bookings"
        ]
      },
      "get": {
        "operationId": "BookingFileController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingFileResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all workspace files for a booking",
        "tags": [
          "Service Bookings"
        ]
      }
    },
    "/api/bookings/{id}/files/{fileId}": {
      "get": {
        "operationId": "BookingFileController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingFileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get booking workspace file details",
        "tags": [
          "Service Bookings"
        ]
      },
      "delete": {
        "operationId": "BookingFileController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a booking workspace file",
        "tags": [
          "Service Bookings"
        ]
      }
    },
...
    "/api/moderation/bookings/{id}": {
      "get": {
        "operationId": "AdminController_getBookingDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get booking detail with related transactions",
        "tags": [
          "Admin Moderation"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gripper API",
    "description": "The comprehensive API for the Gripper platform backend.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.gripper.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BookingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "pending_deposit",
              "confirmed",
              "in_progress",
              "delivered",
              "revision_requested",
              "pending_completion_payment",
              "completed",
              "cancelled",
              "disputed"
            ]
          },
          "scheduledDate": {
            "format": "date-time",
            "type": "string"
          },
          "scheduledTime": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "requirements": {
            "type": "object"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "confirmedAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string"
          },
          "cancellationReason": {
            "type": "string"
          },
          "cancelledBy": {
            "type": "string"
          },
          "service": {
            "type": "object"
          },
          "customer": {
            "type": "object"
          },
          "provider": {
            "type": "object"
          },
          "referenceCode": {
            "type": "string"
          },
          "depositAmount": {
            "type": "number"
          },
          "platformFee": {
            "type": "number"
          },
          "depositPaid": {
            "type": "boolean"
          },
          "finalPaymentPaid": {
            "type": "boolean"
          },
          "correctionsUsed": {
            "type": "number"
          },
          "correctionsLimit": {
            "type": "number"
          },
          "correctionFee": {
            "type": "number"
          },
          "customerApproved": {
            "type": "boolean"
          },
          "customerApprovedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "providerId",
          "serviceId",
          "status",
          "amount",
          "currency",
          "depositPaid",
          "finalPaymentPaid",
          "correctionsUsed",
          "correctionsLimit",
          "correctionFee",
          "customerApproved",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "example": "uuid-of-service"
          },
          "scheduledDate": {
            "format": "date-time",
            "type": "string",
            "example": "2023-12-25T10:00:00Z"
          },
          "scheduledTime": {
            "type": "string",
            "example": "10:00 AM"
          },
          "notes": {
            "type": "string",
            "example": "Please arrive 10 minutes early."
          },
          "location": {
            "type": "string",
            "example": "123 Main St, New York, NY"
          },
          "requirements": {
            "type": "object",
            "example": {
              "color": "blue"
            }
          }
        },
        "required": [
          "serviceId"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "scheduledDate": {
            "format": "date-time",
            "type": "string",
            "example": "2023-12-25T11:00:00Z"
          },
          "scheduledTime": {
            "type": "string",
            "example": "11:00 AM"
          },
          "notes": {
            "type": "string",
            "example": "Updated notes."
          },
          "location": {
            "type": "string",
            "example": "456 Second St, New York, NY"
          }
        }
      },
...
      "RequestCorrectionDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Please adjust spacing and update the hero image."
          },
          "attachments": {
            "example": [
              "https://cdn.example.com/screenshot.png"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "description"
        ]
      },
      "CancelBookingDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Conflict of schedule"
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "BookingFileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bookingId": {
            "type": "string"
          },
          "uploadedBy": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "url": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "bookingId",
          "uploadedBy",
          "filename",
          "originalName",
          "mimeType",
          "size",
          "url",
          "fileType",
          "createdAt"
        ]
      },
...
    }
  }
}